+++ src/zope/intid/__init__.py
@@ -36,7 +36,18 @@ from zope.intid.interfaces import IIntIds, IIntIdEvent
 from zope.intid.interfaces import IntIdAddedEvent, IntIdRemovedEvent
 from zope.intid.interfaces import IntIdMissingError, IntIdsCorruptedError, ObjectMissingError
 
-
+try:
+    # POSKeyError is a subclass of KeyError; in the cases where we
+    # catch KeyError for an item missing from a BTree, we still
+    # want to propagate this exception that indicates a corrupt database
+    # (as opposed to a corrupt IntIds)
+    from ZODB.POSException import POSKeyError as _POSKeyError
+except ImportError: # pragma: no cover (we run tests with ZODB installed)
+    # In practice, ZODB will probably be installed. But if not,
+    # then POSKeyError can never be generated, so use a unique
+    # exception that we'll never catch.
+    class _POSKeyError(BaseException):
+        pass
 
 @implementer(IIntIds, IContained)
 class IntIds(Persistent):
@@ -72,11 +83,9 @@ class IntIds(Persistent):
     def getObject(self, id):
         try:
             return self.refs[id]()
+        except _POSKeyError:
+            raise
         except KeyError:
-            # In theory, if the ZODB and/or our self.ids BTree is
-            # corrupted, this could raise
-            # ZODB.POSException.POSKeyError. We used to propagate that
-            # but now we transform it.
             raise ObjectMissingError(id)
 
     def queryObject(self, id, default=None):
@@ -93,17 +102,16 @@ class IntIds(Persistent):
 
         try:
             return self.ids[key]
+        except _POSKeyError:
+            raise
         except KeyError:
-            # In theory, if the ZODB and/or our self.ids BTree is
-            # corrupted, this could raise
-            # ZODB.POSException.POSKeyError, which we should probably
-            # let propagate. But since that's a KeyError, we've always
-            # caught it and transformed the error message and type.
             raise IntIdMissingError(ob)
 
     def queryId(self, ob, default=None):
         try:
             return self.getId(ob)
+        except _POSKeyError:
+            raise
         except KeyError:
             return default
 
@@ -146,26 +154,27 @@ class IntIds(Persistent):
         if key is None:
             return
 
-        # In theory, any of the KeyErrors we're catching here could be
-        # a ZODB POSKeyError if the ZODB and our BTrees are corrupt.
-        # We used to let those propagate (where they would probably be
-        # ignored, see removeIntIdSubscriber), but now we transform
-        # them and ignore that possibility because the chances of that
-        # happening are (probably) extremely remote.
-
         try:
             uid = self.ids[key]
+        except _POSKeyError:
+            raise
         except KeyError:
             raise IntIdMissingError(ob)
 
         try:
             del self.refs[uid]
+        except _POSKeyError:
+            raise
         except KeyError:
             # It was in self.ids, but not self.refs. Something is corrupted.
             # We've always let this KeyError propagate, before cleaning up self.ids,
             # meaning that getId(ob) will continue to work, but getObject(uid) will not.
             raise IntIdsCorruptedError(ob, uid)
         del self.ids[key]
+
+def _utilities_and_key(ob):
+    utilities = tuple(getAllUtilitiesRegisteredFor(IIntIds))
+    return utilities, IKeyReference(ob, None) if utilities else None
 
 @adapter(ILocation, IObjectRemovedEvent)
 def removeIntIdSubscriber(ob, event):
@@ -174,19 +183,18 @@ def removeIntIdSubscriber(ob, event):
     Removes the unique ids registered for the object in all the unique
     id utilities.
     """
-    utilities = tuple(getAllUtilitiesRegisteredFor(IIntIds))
-    if utilities:
-        key = IKeyReference(ob, None)
-        # Register only objects that adapt to key reference
-        if key is not None:
-            # Notify the catalogs that this object is about to be removed.
-            notify(IntIdRemovedEvent(ob, event))
-            for utility in utilities:
-                try:
-                    utility.unregister(key)
-                except KeyError:
-                    # Silently ignoring corruption here
-                    pass
+    utilities, key = _utilities_and_key(ob)
+    if not utilities or key is None:
+        # Unregister only objects that adapt to key reference
+        return
+    # Notify the catalogs that this object is about to be removed.
+    notify(IntIdRemovedEvent(ob, event))
+    for utility in utilities:
+        try:
+            utility.unregister(key)
+        except KeyError:
+            # Silently ignoring all kinds corruption here
+            pass
 
 @adapter(ILocation, IObjectAddedEvent)
 def addIntIdSubscriber(ob, event):
@@ -195,16 +203,15 @@ def addIntIdSubscriber(ob, event):
     Registers the object added in all unique id utilities and fires
     an event for the catalogs.
     """
-    utilities = tuple(getAllUtilitiesRegisteredFor(IIntIds))
-    if utilities: # assert that there are any utilites
-        key = IKeyReference(ob, None)
+    utilities, key = _utilities_and_key(ob)
+    if not utilities or key is None:
         # Register only objects that adapt to key reference
-        if key is not None:
-            idmap = {}
-            for utility in utilities:
-                idmap[utility] = utility.register(key)
-            # Notify the catalogs that this object was added.
-            notify(IntIdAddedEvent(ob, event, idmap))
+        return
+    idmap = {}
+    for utility in utilities:
+        idmap[utility] = utility.register(key)
+    # Notify the catalogs that this object was added.
+    notify(IntIdAddedEvent(ob, event, idmap))
 
 @adapter(IIntIdEvent)
 def intIdEventNotify(event):